KILL ME NOW

+++ 0x0A-configuration_management/2-execute_a_command.pp
@@ -1,5 +1,7 @@
-# Executes a bash command
-exec { 'kill':
-  command => 'pkill -f killmenow',
+# Define an exec resource to kill the process named "killmenow"
+exec { 'killmenow':
+  command     => 'pkill -f killmenow',
+  refreshonly => true, # Only run the command when notified
+  onlyif      => 'pgrep killmenow', # Check if the process is running before attempting to kill
   path    => ['/usr/bin', '/usr/sbin']
 }
